Unify try-name
